# tests/fixtures/http/openapi.yaml
openapi: 3.1.0
info:
  title: HTTP fixture
  version: 1.0.0
security:
  - sessionCookie: []
paths:
  /health:
    get:
      security: []
      operationId: health
      responses:
        "204":
          description: Ready
  /widgets/{id}:
    get:
      operationId: getWidget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 10
      responses:
        "200":
          description: Widget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Widget"
        "400":
          description: Invalid widget request
    post:
      operationId: updateWidget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: wait
          in: query
          required: true
          schema:
            type: string
            enum:
              - "0"
        - name: X-Widget-Trace
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WidgetUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Widget"
          links:
            readWidget:
              operationId: getWidget
              parameters:
                id: $response.body#/id
        "400":
          description: Invalid widget request
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  schemas:
    WidgetUpdate:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 128
    Widget:
      type: object
      additionalProperties: false
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
          maxLength: 128